Fix the error of styling not showing
The problem of the CSS not showing on the page
has been fixed. I forgot to add the link to the
style page in the html file. The link has been
added and the styling is now showing properly.

The font style has been added to the page.
Background and font colors have been added to
the hero section, header, and footer.