Begin to style the hero section and image
I have added styling to the hero section with flex
to match the spacing with the header. I have also
added an image folder to the directory and linked
an image to the html.

I still need to style the image along with figure
out the spacing of the elements within the hero
section.